refactor: Colocate unit plans and files
